When testing saveValueToKey log that key from all nodes after
As a way of seeing which nodes got the key written to them, which ones
didnt and what the value is for the key.

This made me realize that I have a mistake with writing to multiple
nodes. The reader can only be read from once and so only one of the
nodes will get the value written to them. So I need to split the reader
into a multireader if I want to send it to be written to multiple nodes